worker/requirements.txt
requests[socks]
PySocks
asyncio
asyncpg
python-dotenv
